+++ CHANGELOG.md
@@ -28,6 +28,7 @@ You can also check the
 - Style
   - Improved vertical spacing between map legend items
   - Fixed the spacing between navigation and header in the /profile view
+  - Regular charts now have consistent margin between the Y axis label and ticks
 
 # [5.2.1] - 2025-01-29